third_party/blink/web_tests/http/tests/cookies/simple-cookies-expired-expected.txt

This test checks that cookies are correctly set using Expires.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that setting a simple cookie works.
PASS cookie is 'test=foobar'.
Check setting a cookie that timed out.
PASS cookie is ''.
PASS successfullyParsed is true

TEST COMPLETE
